+++ planner_cspace/test/src/test_cyclic_vec.cpp
@@ -104,7 +104,7 @@ TEST(CyclicVec, LengthInt)
 
   ASSERT_EQ(v.sqlen(), 25);
   ASSERT_EQ(v.len(), 5.0);
-  ASSERT_EQ(v.norm(), sqrtf(50));
+  ASSERT_EQ(v.norm(), std::sqrt(50));
 }
 
 TEST(CyclicVec, LengthFloat)
@@ -114,7 +114,7 @@ TEST(CyclicVec, LengthFloat)
 
   ASSERT_EQ(v.sqlen(), 25.0);
   ASSERT_EQ(v.len(), 5.0);
-  ASSERT_EQ(v.norm(), sqrtf(50.0));
+  ASSERT_EQ(v.norm(), std::sqrt(50.0));
 }
 
 TEST(CyclicVec, Cycle)
